# mp-app/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    PackageCreate:
      additionalProperties: false
      properties:
        duration_days:
          maximum: 3650.0
          minimum: 1.0
          title: Duration Days
          type: integer
        name:
          maxLength: 80
          minLength: 1
          title: Name
          type: string
        price_rial:
          minimum: 0.0
          title: Price Rial
          type: integer
      required:
      - name
      - duration_days
      - price_rial
      title: PackageCreate
      type: object
    PaymentCreate:
      additionalProperties: false
      properties:
        amount_rial:
          minimum: 0.0
          title: Amount Rial
          type: integer
        member_id:
          title: Member Id
          type: integer
        method:
          default: cash
          pattern: ^(cash|card|transfer|pos)$
          title: Method
          type: string
        package_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Package Id
      required:
      - member_id
      - amount_rial
      title: PaymentCreate
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  description: Local-first Gym OS core API (offline-capable).
  title: Muscle Paradise Core
  version: 0.20.0
  x-generated-by: app.export_openapi
openapi: 3.1.0
paths:
...
  /api/v1/packages:
    get:
      operationId: list_packages_api_v1_packages_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                title: Response List Packages Api V1 Packages Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Active membership packages
      tags:
      - payments
    post:
      operationId: create_package_api_v1_packages_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Package Api V1 Packages Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Define a package (finance only)
      tags:
      - payments
  /api/v1/payments:
    post:
      operationId: create_payment_api_v1_payments_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Payment Api V1 Payments Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Record a payment (cash entry at the desk)
      tags:
      - payments
  /api/v1/payments/{payment_id}/receipt:
    get:
      operationId: receipt_api_v1_payments__payment_id__receipt_get
      parameters:
      - in: path
        name: payment_id
        required: true
        schema:
          title: Payment Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Receipt PDF
      tags:
      - payments
  /api/v1/payments/{payment_id}/void:
    post:
      operationId: void_api_v1_payments__payment_id__void_post
      parameters:
      - in: path
        name: payment_id
        required: true
        schema:
          title: Payment Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Void Api V1 Payments  Payment Id  Void Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Void (finance only, audited)
      tags:
      - payments
...